backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
sarvamai
google-generativeai
langdetect
aiofiles
